move file explorer widget to shared
