--- apps/file-explorer-e2e/cypress.config.ts
@@ -1,6 +1,0 @@
-import { defineConfig } from 'cypress';
-import { nxE2EStorybookPreset } from '@nrwl/storybook/presets/cypress';
-
-export default defineConfig({
-  e2e: nxE2EStorybookPreset(__dirname),
-});
