--- apps/file-explorer-e2e/project.json
@@ -1,30 +1,0 @@
-{
-  "name": "file-explorer-e2e",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "apps/file-explorer-e2e/src",
-  "projectType": "application",
-  "targets": {
-    "e2e": {
-      "executor": "@nrwl/cypress:cypress",
-      "options": {
-        "cypressConfig": "apps/file-explorer-e2e/cypress.config.ts",
-        "devServerTarget": "file-explorer:storybook",
-        "testingType": "e2e"
-      },
-      "configurations": {
-        "ci": {
-          "devServerTarget": "file-explorer:storybook:ci"
-        }
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["apps/file-explorer-e2e/**/*.{js,ts}"]
-      }
-    }
-  },
-  "tags": [],
-  "implicitDependencies": ["file-explorer"]
-}
