--- apps/file-explorer-e2e/src/e2e/components/file-explorer-widget/file-explorer-widget.component.cy.ts
@@ -1,8 +1,0 @@
-describe('file-explorer', () => {
-  beforeEach(() =>
-    cy.visit('/iframe.html?id=fileexplorerwidgetcomponent--primary')
-  );
-  it('should render the component', () => {
-    cy.get('oort-front-file-explorer-widget').should('exist');
-  });
-});
