--- libs/file-explorer/.storybook/preview.js
@@ -1,10 +1,0 @@
-import '@angular/localize/init';
-// import { setCompodocJson } from '@storybook/addon-docs/angular';
-// import docJson from '../documentation.json';
-// setCompodocJson(docJson);
-import { library } from '@fortawesome/fontawesome-svg-core';
-import { fas } from '@fortawesome/free-solid-svg-icons';
-import { fab } from '@fortawesome/free-brands-svg-icons';
-
-// Add font awesome icons
-library.add(fas, fab);
