--- libs/file-explorer/README.md
@@ -1,23 +1,0 @@
-# file-explorer
-
-This library was generated with [Nx](https://nx.dev).
-
-## Run storybook
-
-`npx nx run file-explorer:storybook`
-
-## Lint
-
-`npx nx run file-explorer:lint`
-
-## Generate new Angular files
-
-Generate a new component:
-`npx nx g c components/<name> --project=file-explorer`
-
-Generate a new standalone component:
-`npx nx g c components/<name> --standalone true --project=file-explorer`
-
-## Running unit tests
-
-Run `nx test file-explorer` to execute the unit tests.
