--- libs/file-explorer/project.json
@@ -1,82 +1,0 @@
-{
-  "name": "file-explorer",
-  "$schema": "../node_modules/nx/schemas/project-schema.json",
-  "projectType": "library",
-  "sourceRoot": "libs/file-explorer/src",
-  "prefix": "oort-front",
-  "targets": {
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "libs/file-explorer/jest.config.ts",
-        "passWithNoTests": true
-      },
-      "configurations": {
-        "ci": {
-          "ci": true,
-          "codeCoverage": true
-        }
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": [
-          "libs/file-explorer/**/*.ts",
-          "libs/file-explorer/**/*.html"
-        ]
-      }
-    },
-    "storybook": {
-      "executor": "@storybook/angular:start-storybook",
-      "options": {
-        "port": 4400,
-        "configDir": "libs/file-explorer/.storybook",
-        "browserTarget": "file-explorer:build-storybook",
-        "compodoc": false,
-        "styles": [
-          "libs/shared/src/lib/style/styles.scss",
-          "libs/styles/src/lib/storybook/storybook.scss",
-          "libs/styles/src/lib/tailwind/tailwind.scss",
-          "libs/styles/src/lib/shared/shared.scss",
-          "libs/styles/src/lib/kendo/theme.scss",
-          "node_modules/survey-core/defaultV2.min.css",
-          "node_modules/survey-core/survey.min.css"
-        ],
-        "stylePreprocessorOptions": {
-          "includePaths": ["libs/styles/src/lib/themes/default"]
-        }
-      },
-      "configurations": {
-        "ci": {
-          "quiet": true
-        }
-      }
-    },
-    "build-storybook": {
-      "executor": "@storybook/angular:build-storybook",
-      "outputs": ["{options.outputDir}"],
-      "options": {
-        "outputDir": "dist/storybook/file-explorer",
-        "configDir": "libs/file-explorer/.storybook",
-        "browserTarget": "file-explorer:build-storybook",
-        "compodoc": false,
-        "styles": ["libs/shared/src/lib/style/styles.scss"],
-        "stylePreprocessorOptions": {
-          "includePaths": [
-            "libs/styles/src/lib/themes/default",
-            "libs/styles/src/lib/kendo"
-          ]
-        }
-      },
-      "configurations": {
-        "ci": {
-          "quiet": true
-        }
-      }
-    }
-  },
-  "tags": []
-}
