--- libs/file-explorer/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/file-explorer.module';
