--- libs/file-explorer/src/lib/file-explorer.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-/**
- * File explorer module.
- */
-@NgModule({
-  imports: [CommonModule],
-})
-export class FileExplorerModule {}
