--- libs/file-explorer/tsconfig.lib.json
@@ -1,19 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
-    "declaration": true,
-    "declarationMap": true,
-    "inlineSources": true,
-    "types": []
-  },
-  "exclude": [
-    "src/test-setup.ts",
-    "src/**/*.spec.ts",
-    "jest.config.ts",
-    "src/**/*.test.ts",
-    "**/*.stories.ts",
-    "**/*.stories.js"
-  ],
-  "include": ["src/**/*.ts"]
-}
