+++ libs/shared/.storybook/preview.js
@@ -2,3 +2,9 @@ import '@angular/localize/init';
 import { setCompodocJson } from '@storybook/addon-docs/angular';
 import docJson from '../documentation.json';
 setCompodocJson(docJson);
+
+// Add font awesome icons
+import { library } from '@fortawesome/fontawesome-svg-core';
+import { fas } from '@fortawesome/free-solid-svg-icons';
+import { fab } from '@fortawesome/free-brands-svg-icons';
+library.add(fas, fab);
