+++ libs/shared/README.md
@@ -2,6 +2,23 @@
 
 This library was generated with [Nx](https://nx.dev).
 
+## Run storybook
+
+`npx nx run shared:storybook`
+
+## Lint
+
+`npx nx run shared:lint`
+
+## Generate new Angular files
+
+Generate a new component:
+`npx nx g c components/<name> --project=shared`
+
+Generate a new standalone component:
+`npx nx g c components/<name> --standalone true --project=shared`
+
+
 ## Running unit tests
 
 Run `nx test shared` to execute the unit tests.
