+++ libs/shared/src/lib/components/file-explorer/file-explorer-list-item/file-explorer-list-item.component.html
