+++ libs/shared/src/lib/components/file-explorer/file-explorer-list-item/file-explorer-list-item.component.ts
@@ -1,8 +1,8 @@
 import { Component, Input } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FileExplorerDocument } from '../../types/file-explorer-document.type';
+import { FileExplorerDocument } from '../types/file-explorer-document.type';
 import { IconModule } from '@oort-front/ui';
-import { FileIconPipe } from '../../pipes/file-icon/file-icon.pipe';
+import { FileIconPipe } from '../../../pipes/file-icon/file-icon.pipe';
 
 /**
  * File explorer list item component.
