+++ libs/shared/src/lib/components/file-explorer/file-explorer-table/file-explorer-table.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { GridModule } from '@progress/kendo-angular-grid';
-import { FileExplorerDocument } from '../../types/file-explorer-document.type';
+import { FileExplorerDocument } from '../types/file-explorer-document.type';
 
 /**
  * File explorer 'grid' view component.
