+++ libs/shared/src/lib/components/file-explorer/file-explorer-widget/file-explorer-widget.component.html
