+++ libs/shared/src/lib/components/file-explorer/file-explorer-widget/file-explorer-widget.component.ts
@@ -3,9 +3,9 @@ import { CommonModule } from '@angular/common';
 import { LayoutModule } from '@progress/kendo-angular-layout';
 import { FileExplorerToolbarComponent } from '../file-explorer-toolbar/file-explorer-toolbar.component';
 import { FileExplorerTableComponent } from '../file-explorer-table/file-explorer-table.component';
-import { fileExplorerView } from '../../types/fie-explorer-view.type';
+import { fileExplorerView } from '../types/file-explorer-view.type';
 import { FileExplorerListComponent } from '../file-explorer-list/file-explorer-list.component';
-import { FileExplorerDocument } from '../../types/file-explorer-document.type';
+import { FileExplorerDocument } from '../types/file-explorer-document.type';
 
 /**
  * File explorer widget component.
