+++ libs/shared/src/lib/components/file-explorer/types/file-explorer-document.type.ts
