+++ libs/shared/src/lib/components/file-explorer/types/file-explorer-view.type.ts
