+++ libs/shared/src/lib/pipes/file-icon/file-icon.pipe.ts
